Make important settings more visible
Make non profile-related settings (such as security settings) more visible
by moving them apart from 'Edit profile' into their own 'Edit settings' page
