+++ app/views/layouts/application.html.erb
@@ -72,6 +72,7 @@
               <div class="header__popup__nav-links">
                 <%= link_to t('.header.dashboard'), dashboard_path, class: "header__nav-link" %>
                 <%= link_to t('.header.edit_profile'), edit_profile_path, class: "header__nav-link" %>
+                <%= link_to t('.header.edit_settings'), edit_settings_path, class: "header__nav-link" %>
                 <%= link_to t('.header.sign_out'), sign_out_path, method: :delete, class: "header__nav-link" %>
               </div>
             <% else %>
