+++ app/views/multifactor_auths/recovery.html.erb
@@ -7,5 +7,5 @@
     <% end %>
   </ul>
   <p><%= t '.note' %></p>
-  <p><%= link_to t('.continue'), edit_profile_path, class: "t-link--arrow" %></p>
+  <p><%= link_to t('.continue'), edit_settings_path, class: "t-link--arrow" %></p>
 </div>
