+++ config/locales/fr.yml
@@ -106,6 +106,7 @@ fr:
       header:
         dashboard: Dashboard
         edit_profile:
+        edit_settings:
         search_gem_html: Recherche de Gems&hellip;
         sign_in: Connexion
         sign_out: Déconnexion
@@ -215,14 +216,9 @@ fr:
       recommended:
       title:
       update:
-  profiles:
+  settings:
     edit:
-      change_avatar:
-      email_awaiting_confirmation: Veuillez confirmer votre nouvelle adresse email %{unconfirmed_email}
-      enter_password: Veuillez entrer le mot de passe de votre compte
-      hide_email: Ne pas afficher l'email dans le profil public
-      optional_twitter_username: Nom d'utilisateur Twitter optionnel. Sera affiché publiquement
-      title: Modification de profil
+      title:
       api_access:
         confirm_reset: Êtes-vous sûr ? Cette action ne peut être annulée.
         credentials_html: 'Si vous voulez utiliser %{gem_commands_link} de la ligne de commande, vous aurez besoin d''un fichier %{gem_credentials_file}, que vous pouvez générer avec la commande suivante :'
@@ -234,10 +230,6 @@ fr:
         link_text: Demandez un nouveau mot de passe ici.
         text_html: Besoin d'un nouveau mot de passe ? Pas de problème. %{link}
         title: Renvoi de mot de passe
-      delete:
-        delete: Supprimer
-        delete_profile: Supprimer le profil
-        warning: La suppression de votre profil est une action irréversible. Cela ne peut pas être annulé en contactant le support. Réfléchissez-bien avant de continuer !
       mfa:
         multifactor_auth: Authentification Multifacteur
         disabled: Vous n'avez pas encore activé l'authentification multifacteur.
@@ -249,6 +241,18 @@ fr:
           disabled:
           ui_only:
           ui_and_api:
+  profiles:
+    edit:
+      change_avatar:
+      email_awaiting_confirmation: Veuillez confirmer votre nouvelle adresse email %{unconfirmed_email}
+      enter_password: Veuillez entrer le mot de passe de votre compte
+      hide_email: Ne pas afficher l'email dans le profil public
+      optional_twitter_username: Nom d'utilisateur Twitter optionnel. Sera affiché publiquement
+      title: Modification de profil
+      delete:
+        delete: Supprimer
+        delete_profile: Supprimer le profil
+        warning: La suppression de votre profil est une action irréversible. Cela ne peut pas être annulé en contactant le support. Réfléchissez-bien avant de continuer !
     delete:
       title: Supprimer le profil
       confirm: Confirmer
